LayoutTests/media/modern-media-controls/tracks-support/tracks-support-click-track-in-panel-expected.txt

Clicking on a track in the tracks panel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.audioTracks[1].enabled is false

Clicking on the tracks button
Clicking on second audio track in tracks panel
Obtained change event on media audio tracks
PASS media.audioTracks[1].enabled is true

PASS successfullyParsed is true

TEST COMPLETE
